split moderation create commands into separate files
