+++ src/main/java/de/nplay/moderationbot/duration/DurationAdapter.java
@@ -3,6 +3,7 @@ package de.nplay.moderationbot.duration;
 import com.github.kaktushose.jda.commands.dispatching.adapter.TypeAdapter;
 import com.github.kaktushose.jda.commands.guice.Implementation;
 import io.github.kaktushose.proteus.mapping.MappingResult;
+import org.jspecify.annotations.NullMarked;
 import org.jspecify.annotations.Nullable;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -23,35 +24,12 @@ public class DurationAdapter implements TypeAdapter<String, Duration> {
             return MappingResult.failure("Die angegebene Dauer ist ungültig. Bitte gib eine gültige Dauer an.");
         }
 
-        var normalizedRaw = source.toUpperCase().replaceAll("\\s+", "");
-
-        var years = extractValue(normalizedRaw, "Y");
-        var months = extractValue(normalizedRaw, "M");
-        var days = extractValue(normalizedRaw, "D");
-        var hours = extractValue(normalizedRaw, "H");
-        var minutes = extractValue(normalizedRaw, "MIN");
-        var seconds = extractValue(normalizedRaw, "S");
-
-        if (years != null) days = (days != null ? days : 0) + years * 365;
-        if (months != null) days = (days != null ? days : 0) + months * 30; // Approximation, as months vary in length
-
-        var parseString = "P" +
-                          (days != null ? days + "D" : "") +
-                          (hours != null || minutes != null || seconds != null ? "T" : "") +
-                          (hours != null ? hours + "H" : "") +
-                          (minutes != null ? minutes + "M" : "") +
-                          (seconds != null ? seconds + "S" : "");
-
-        try {
-            return MappingResult.lossless(Duration.parse(parseString));
-        } catch (DateTimeParseException e) {
-            log.warn("User provided invalid duration: {}", source);
-            return MappingResult.failure("Die angegebene Dauer ist ungültig. Bitte gib eine gültige Dauer an.");
-        }
+        return parse(source)
+                .map(it -> (MappingResult<Duration>) MappingResult.lossless(it))
+                .orElse(MappingResult.failure("Die angegebene Dauer ist ungültig. Bitte gib eine gültige Dauer an."));
     }
 
-    // TODO refactor
-    public static Optional<Duration> parse(@Nullable String raw) {
+    public Optional<Duration> parse(@Nullable String raw) {
         if (raw == null || raw.isBlank()) {
             return Optional.empty();
         }
@@ -83,7 +61,8 @@ public class DurationAdapter implements TypeAdapter<String, Duration> {
         }
     }
 
-    private static Integer extractValue(String input, String unit) {
+    @Nullable
+    private Integer extractValue(String input, String unit) {
         var pattern = "(?i)(\\d+)" + (unit.equals("MIN") ? "MIN" : unit + "(?!IN)");
         var matcher = Pattern.compile(pattern).matcher(input);
         return matcher.find() ? Integer.parseInt(matcher.group(1)) : null;
