+++ src/main/java/module-info.java
@@ -21,4 +21,5 @@ module moderationbot.main {
     requires sadu.sadu.postgresql.main;
     requires sadu.sadu.queries.main;
     requires sadu.sadu.updater.main;
+    requires moderationbot.main;
 }
