+++ src/main/resources/default_de.ftl
@@ -35,69 +35,62 @@ permissions-manage-role-options-role-description = Die Rolle, dessen Berechtigun
 
 # Moderation
 ## General
-paragraph-reference = Welcher Regel-Paragraph ist verletzt worden / soll referenziert werden?
-message-link = Link zu einer Nachricht, die referenziert werden soll.
-prune-duration = Für wie viele Tage in der Vergangenheit sollen Nachrichten dieses Users gelöscht werden?
+-link = Nachrichtenlink
+-link-description = Link zu einer Nachricht, die referenziert werden soll.
+-prune-duration = Für wie viele Tage in der Vergangenheit sollen Nachrichten dieses Users gelöscht werden?.
+-del-days = del_days
+-del-days-description = Für wie viele Tage in der Vergangenheit sollen Nachrichten dieses Users gelöscht werden?
 
-mod-name
-mod-description
+reason-title = Begründung angeben
+reason-field = Begründung der Moderationshandlung
+reason-duration-title = Begründung und Dauer angeben
+duration-field = Dauer der Moderationshandlung
+invalid-duration = Die angegebene Dauer ist ungültig. Bitte gib eine gültige Dauer an.
+invalid-duration-limit = Die angegebene Dauer ist zu lang. Bitte gib eine Dauer von maximal 28 Tagen an.
 
 ## Warn
-warn-target = Der Benutzer, der verwarnt werden soll.
-mod-warn-name
-mod-warn-description
-mod-warn-options-target-name
-mod-warn-options-target-description
-mod-warn-options-paragraph-name
-mod-warn-options-paragraph-description
-mod-warn-options-message_link-name
-mod-warn-options-message_link-description
+mod-warn-description = Verwarnt einen Benutzer
+mod-warn-options-target-name = { -target }
+mod-warn-options-target-description = Der Benutzer, der verwarnt werden soll.
+mod-warn-options-paragraph-name = { -paragraph }
+mod-warn-options-paragraph-description = { -paragraph-description }
+mod-warn-options-message_link-name = { -link }
+mod-warn-options-message_link-description = { -link-description }
 
 ## Timeout
-timeout-target = Der Benutzer, den in den Timeout versetzt werden soll.
-timeout-length = Für wie lange der Timeout andauern soll (max. 28 Tage).
-
-mod-timeout-name
-mod-timeout-description
-mod-timeout-options-target-name
-mod-timeout-options-target-description
-mod-timeout-options-until-name
-mod-timeout-options-until-description
-mod-timeout-options-paragraph-name
-mod-timeout-options-paragraph-description
-mod-timeout-options-message_link-name
-mod-timeout-options-message_link-description
+mod-timeout-description = Versetzt einen Benutzer in den Timeout
+mod-timeout-options-target-name = { -target }
+mod-timeout-options-target-description = Der Benutzer, den in den Timeout versetzt werden soll.
+mod-timeout-options-until-name = { -dauer }
+mod-timeout-options-until-description = Für wie lange der Timeout andauern soll (max. 28 Tage).
+mod-timeout-options-paragraph-name = { -paragraph }
+mod-timeout-options-paragraph-description = { -paragraph-description }
+mod-timeout-options-message_link-name = { -link }
+mod-timeout-options-message_link-description = { -link-description }
 
 ## Kick
-kick-target = Der Benutzer, der gekickt werden soll.
-
-mod-kick-name
-mod-kick-description
-mod-kick-options-target-name
-mod-kick-options-target-description
-mod-kick-options-paragraph-name
-mod-kick-options-paragraph-description
-mod-kick-options-del_days-name
-mod-kick-options-del_days-description
-mod-kick-options-message_link-name
-mod-kick-options-message_link-description
+mod-kick-description = Kickt einen Benutzer vom Server
+mod-kick-options-target-name = { -target }
+mod-kick-options-target-description = Der Benutzer, der gekickt werden soll.
+mod-kick-options-paragraph-name  = { -paragraph }
+mod-kick-options-paragraph-description = { -paragraph-description }
+mod-kick-options-del_days-name = { -del-days }
+mod-kick-options-del_days-description = { -del-days-description }
+mod-kick-options-message_link-name = { -link }
+mod-kick-options-message_link-description = { -link-description }
 
 ## Ban
-ban-target = Der Benutzer, der gebannt werden soll.
-ban-duration = Für wie lange der Ban andauern soll.
-
-mod-ban-name
-mod-ban-description
-mod-ban-options-target-name
-mod-ban-options-target-description
-mod-ban-options-until-name
-mod-ban-options-until-description
-mod-ban-options-del_days-name
-mod-ban-options-del_days-description
-mod-ban-options-paragraph-name
-mod-ban-options-paragraph-description
-mod-ban-options-message_link-name
-mod-ban-options-message_link-description
+mod-ban-description = Bannt einen Benutzer vom Server
+mod-ban-options-target-name = { -target }
+mod-ban-options-target-description = Der Benutzer, der gebannt werden soll.
+mod-ban-options-until-name = { -dauer }
+mod-ban-options-until-description = Für wie lange der Ban andauern soll.
+mod-ban-options-del_days-name = { -del-days }
+mod-ban-options-del_days-description = { -del-days-description }
+mod-ban-options-paragraph-name = { -paragraph }
+mod-ban-options-paragraph-description = { -paragraph-description }
+mod-ban-options-message_link-name = { -link }
+mod-ban-options-message_link-description = { -link-description }
 
 ## Revert
 revert-act = Die ID der Moderationshandlung, die aufgehoben werden soll.
